fix spelling of behavior and start fixing bdd steps
